requirements.txt
playwright>=1.40.0
pandas>=2.1.0
pyyaml>=6.0
pytest>=7.4.0
pytest-playwright>=0.4.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
transformers>=4.30.0
torch
tqdm
